+++ example/android/app/src/main/java/com/tundralabs/flutterttsexample/MainActivity.java
@@ -1,4 +1,4 @@
-package com.eyedeadevelopers.flutterttsexample;
+package com.tundralabs.flutterttsexample;
 
 import android.os.Bundle;