+++ lib/flutter_tts.dart
@@ -5,9 +5,11 @@ import 'package:flutter/services.dart';
 
 typedef void ErrorHandler(dynamic message);
 
+// Provides Platform specific TTS services (Android: TextToSpeech, IOS: AVSpeechSynthesizer)
 class FlutterTts {
   static const MethodChannel _channel = const MethodChannel('flutter_tts');
 
+  VoidCallback initHandler;
   VoidCallback startHandler;
   VoidCallback completionHandler;
   ErrorHandler errorHandler;
@@ -16,29 +18,54 @@ class FlutterTts {
     _channel.setMethodCallHandler(platformCallHandler);
   }
 
+  /// [Future] which invokes the platform specific method for speaking
   Future<dynamic> speak(String text) => _channel.invokeMethod('speak', text);
 
+  /// [Future] which invokes the platform specific method for setLanguage
   Future<dynamic> setLanguage(String language) =>
       _channel.invokeMethod('setLanguage', language);
 
+  /// [Future] which invokes the platform specific method for setSpeechRate
+  /// Allowed values are in the range from 0.0 (silent) to 1.0 (loudest)
   Future<dynamic> setSpeechRate(double rate) =>
       _channel.invokeMethod('setSpeechRate', rate);
 
+  /// [Future] which invokes the platform specific method for setVolume
+  /// Allowed values are in the range from 0.0 (silent) to 1.0 (loudest)
   Future<dynamic> setVolume(double volume) =>
       _channel.invokeMethod('setVolume', volume);
 
+  /// [Future] which invokes the platform specific method for setPitch
+  /// 1.0 is default and ranges from .5 to 2.0
   Future<dynamic> setPitch(double pitch) =>
       _channel.invokeMethod('setPitch', pitch);
 
+  /// [Future] which invokes the platform specific method for setVoice
+  /// ***Android supported only***
+  Future<dynamic> setVoice(String voice) =>
+      _channel.invokeMethod('setVoice', voice);
+
+  /// [Future] which invokes the platform specific method for stop
   Future<dynamic> stop() => _channel.invokeMethod('stop');
 
-  Future<dynamic> pause() => _channel.invokeMethod('pause');
-
-  Future<List<dynamic>> get getLanguages async {
-    final List<dynamic> languages = await _channel.invokeMethod('getLanguages');
+  /// [Future] which invokes the platform specific method for getLanguages
+  /// Android issues with API 21 & 22
+  /// Returns a list of available languages
+  Future<dynamic> get getLanguages async {
+    final languages = await _channel.invokeMethod('getLanguages');
     return languages;
   }
 
+  /// [Future] which invokes the platform specific method for getVoices
+  /// ***Android supported only ***
+  /// Returns a `List` of voice names
+  Future<dynamic> get getVoices async {
+    final voices = await _channel.invokeMethod('getVoices');
+    return voices;
+  }
+
+  /// [Future] which invokes the platform specific method for isLanguageAvailable
+  /// Returns `true` or `false`
   Future<dynamic> isLanguageAvailable(String language) => _channel.invokeMethod(
       'isLanguageAvailable', <String, Object>{'language': language});
 
@@ -54,8 +81,18 @@ class FlutterTts {
     errorHandler = handler;
   }
 
+  void ttsInitHandler(VoidCallback handler) {
+    initHandler = handler;
+  }
+
+  /// Platform listeners
   Future platformCallHandler(MethodCall call) async {
     switch (call.method) {
+      case "tts.init":
+        if (initHandler != null) {
+          initHandler();
+        }
+        break;
       case "speak.onStart":
         if (startHandler != null) {
           startHandler();